update: cleanup types
