+++ src/components/contextual/pages/pools/types/index.ts
@@ -1,0 +1,16 @@
+export type UserGuageShare = {
+  id: string;
+  gauge: {
+    poolId: string;
+  };
+  balance: string;
+};
+
+export type LiquidityGauge = {
+  poolId: string;
+};
+
+export type UserGuageSharesResponse = {
+  gaugeShares: UserGuageShare[];
+  liquidityGauges: LiquidityGauge[];
+};
